refactor(components): reorganize components and fix imports
