--- components/Component/index.js
@@ -1,13 +1,0 @@
-import AttrDesc from './AttrDesc.vue'
-import AttrTitle from './AttrTitle.vue'
-import ComponentCodeBlock from './CodeBlock.vue'
-import ComponentTitle from './Title.vue'
-import ComponentGrid from './Grid.vue'
-
-export {
-  AttrDesc,
-  AttrTitle,
-  ComponentCodeBlock,
-  ComponentTitle,
-  ComponentGrid
-}
