+++ components/Docs/index.js
@@ -1,0 +1,5 @@
+export { default as AttrDesc } from './AttrDesc.vue'
+export { default as AttrTitle } from './AttrTitle.vue'
+export { default as CalloutCard } from './CalloutCard.vue'
+export { default as CodeBlock } from './CodeBlock.vue'
+export { default as Title } from './Title.vue'
