+++ components/Home/HeaderSection.vue
@@ -16,7 +16,7 @@
 
 <script>
 import UiNav from '~/components/Nav'
-import UiButton from '~/components/Component/Button.vue'
+import UiButton from '~/components/Button.vue'
 import { SpacePeople, Stars } from '~/assets'
 
 import { TipeTextInput, TipeField, TipeTextarea, TipeButton } from '@tipe/ui'
