+++ pages/component/select-button.vue
@@ -1,14 +1,14 @@
 <template>
   <component-grid>
-    <component-title title="Select Button" />
-    <component-code-block sourcecode="import { TipeSelect } from '@tipe/ui'"/>
+    <title title="Select Button" />
+    <code-block sourcecode="import { TipeSelect } from '@tipe/ui'"/>
     <attr-title title="Size"/>
     <tipe-card class="flex">
       <tipe-select type="button" :options="options" placeholder="Select a Role" size="small"/>
       <tipe-select type="button" :options="options" placeholder="Select a Role" size="medium"/>
       <tipe-select type="button" :options="options" placeholder="Select a Role" size="large"/>
     </tipe-card>
-    <component-code-block :sourcecode="sizeSourcecode"/>
+    <code-block :sourcecode="sizeSourcecode"/>
     <attr-title title="Color"/>
     <tipe-card class="flex">
       <tipe-select type="button" :options="options" placeholder="Select a Role" color="primary"/>
@@ -17,26 +17,24 @@
       <tipe-select type="button" :options="options" placeholder="Select a Role" color="warning"/>
       <tipe-select type="button" :options="options" placeholder="Select a Role" color="danger"/>
     </tipe-card>
-    <component-code-block :sourcecode="colorSourcecode"/>
+    <code-block :sourcecode="colorSourcecode"/>
   </component-grid>
 </template>
 
 <script>
 import { TipeCard, TipeSelect } from '@tipe/ui'
-import ComponentTitle from '~/components/Component/Title.vue'
-import ComponentGrid from '~/components/Component/Grid.vue'
-import ComponentCodeBlock from '~/components/Component/CodeBlock.vue'
-import AttrTitle from '~/components/Component/AttrTitle.vue'
+import { Title, AttrTitle, CodeBlock } from '~/components/Docs'
+import ComponentGrid from '~/components/Grid.vue'
 
 export default {
   layout: 'docs',
   components: {
     TipeCard,
     TipeSelect,
-    ComponentTitle,
+    Title,
     ComponentGrid,
     AttrTitle,
-    ComponentCodeBlock
+    CodeBlock
   },
   data() {
     return {
