+++ pages/component/select.vue
@@ -1,40 +1,38 @@
 <template>
   <component-grid>
-    <component-title title="Select" />
-    <component-code-block sourcecode="import { TipeSelect } from '@tipe/ui'"/>
+    <title title="Select" />
+    <code-block sourcecode="import { TipeSelect } from '@tipe/ui'"/>
     <attr-title title="Size"/>
     <tipe-card class="flex">
       <tipe-select :options="options" placeholder="Select a Role" size="small"/>
       <tipe-select :options="options" placeholder="Select a Role" size="medium"/>
       <tipe-select :options="options" placeholder="Select a Role" size="large"/>
     </tipe-card>
-    <component-code-block :sourcecode="sizeSourcecode"/>
+    <code-block :sourcecode="sizeSourcecode"/>
     <attr-title title="Status"/>
     <tipe-card class="flex">
       <tipe-select :options="options" placeholder="Select a Role" status="success"/>
       <tipe-select :options="options" placeholder="Select a Role" status="warning"/>
       <tipe-select :options="options" placeholder="Select a Role" status="error"/>
     </tipe-card>
-    <component-code-block :sourcecode="sizeSourcecode"/>
+    <code-block :sourcecode="sizeSourcecode"/>
   </component-grid>
 </template>
 
 <script>
 import { TipeCard, TipeSelect } from '@tipe/ui'
-import ComponentTitle from '~/components/Component/Title.vue'
-import ComponentGrid from '~/components/Component/Grid.vue'
-import ComponentCodeBlock from '~/components/Component/CodeBlock.vue'
-import AttrTitle from '~/components/Component/AttrTitle.vue'
+import { Title, AttrTitle, CodeBlock } from '~/components/Docs'
+import ComponentGrid from '~/components/Grid.vue'
 
 export default {
   layout: 'docs',
   components: {
     TipeCard,
     TipeSelect,
-    ComponentTitle,
+    Title,
     ComponentGrid,
     AttrTitle,
-    ComponentCodeBlock
+    CodeBlock
   },
   data() {
     return {
