+++ pages/component/switch.vue
@@ -1,45 +1,43 @@
 <template>
   <component-grid>
-    <component-title title="Switch" />
-    <component-code-block sourcecode="import { TipeSwitch } from '@tipe/ui'"/>
+    <title title="Switch" />
+    <code-block sourcecode="import { TipeSwitch } from '@tipe/ui'"/>
     <attr-title title="Size"/>
     <tipe-card class="flex">
       <tipe-switch label="small" size="small"/>
       <tipe-switch label="medium" size="medium"/>
       <tipe-switch label="large" size="large"/>
     </tipe-card>
-    <component-code-block :sourcecode="sizeSourcecode"/>
+    <code-block :sourcecode="sizeSourcecode"/>
     <attr-title title="Status"/>
     <tipe-card id="status">
         <tipe-switch label="success" status="success" size="large"/>
         <tipe-switch label="warning" status="warning" size="large"/>
         <tipe-switch label="error" status="error" size="large"/>
     </tipe-card>
-    <component-code-block :sourcecode="statusSourcecode"/>
+    <code-block :sourcecode="statusSourcecode"/>
     <attr-title title="States"/>
     <tipe-card id="states">
         <tipe-switch label="disabled" size="large" disabled/>
         <tipe-switch label="waiting" size="large" waiting/>
     </tipe-card>
-    <component-code-block :sourcecode="stateSourcecode"/>
+    <code-block :sourcecode="stateSourcecode"/>
   </component-grid>
 </template>
 
 <script>
 import { TipeCard, TipeSwitch, TipeField } from '@tipe/ui'
-import ComponentTitle from '~/components/Component/Title.vue'
-import ComponentGrid from '~/components/Component/Grid.vue'
-import ComponentCodeBlock from '~/components/Component/CodeBlock.vue'
-import AttrTitle from '~/components/Component/AttrTitle.vue'
+import { Title, AttrTitle, CodeBlock } from '~/components/Docs'
+import ComponentGrid from '~/components/Grid.vue'
 
 export default {
   layout: 'docs',
   components: {
     TipeCard,
     TipeSwitch,
-    ComponentTitle,
+    Title,
     ComponentGrid,
-    ComponentCodeBlock,
+    CodeBlock,
     AttrTitle,
     TipeField
   },
